\name{UtilityFiles}
\alias{exportDataFrame}
\alias{exportVector}
\alias{importDataFrame}
\alias{guessExcelColType}
\alias{as.chron.excelDate}
\alias{as.excelData.chron}
\title{Excel file utilities}
\description{Read from and write to MS-Excel from R.}
%\usage{}
%\arguments{
%  \item{...}{}
%}
\details{
  These functions are internal to the package and should not be called
  directly by the user.
}
\author{Duncan Temple Lang <duncan@research.bell-labs.com>.  See RDCOMClient.}
\keyword{IO}
